refactor: PathX is now a partial class
This way, all overloads of a method can be placed in the same file.
